+++ lib/ayeaye/connectors/__init__.py
@@ -9,7 +9,7 @@ from .json_connector import JsonConnector
 from .kafka_connector import KafkaConnector
 from .ndjson_connector import NdjsonConnector
 from .parquet_connector import ParquetConnector
-from .restful_connector import RestfulConnector
+from .restful_connector import RestfulConnector, RestfulConnectorConnectionError
 from .sqlalchemy_database import SqlAlchemyDatabaseConnector
 from .uncooked_connector import UncookedConnector